% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wacc.R
\name{betas}
\alias{betas}
\title{Downloads and tidies Aswath Damodaran's industry Beta data}
\usage{
betas()
}
\description{
Downloads and tidies Aswath Damodaran's Beta data for various sectors of US industry.
}
\details{
Returns a data frame with the following variables:
\itemize{
 \item{\code{Industry: }}{economic sector.}
 \item{\code{Number_Firms: }}{number of companies in the sector.}
 \item{\code{Av_Unlevered_Beta: }}{Average Unlevered Beta.}
 \item{\code{Av_Levered_Beta: }}{Average Levered Beta.}
 \item{\code{Av_Corr_Market: }}{Average Correlation with the market.}
 \item{\code{Total_Unlevered_Beta}}
 \item{\code{Total_Levered_Beta}}
}
}
\examples{
\dontrun{
Beta <- betas()
}
}
